requirements.txt
pdfplumber
pandas 
cohere 
streamlit 
python-dotenv
Pypdf2